Pack-fraction monster system: rooms define monsters as % of a pack
Monster counts in dungeon rooms are now pack fractions (0.0-1.0) instead of
absolute counts. A fraction of 1.0 = a full challenging pack; fractions in a
room must sum to 1.0. At resolution time, fraction × randint(pack_min,pack_max)
gives the count, then difficulty tier scaling applies on top.

- Add pack_min/pack_max to all non-boss monsters in monsters.json
- Load pack sizes into game.monster_stats via validation.py
- Convert all 120+ dungeon room files from absolute counts to fractions
- Update _resolve_dynamic_monsters() to use pack fraction → count conversion
- AI rooms now output monster_groups with fractions (not monster_placements)
- Auto-patcher normalizes AI fractions to sum to 1.0
- Update AI prompts to explain pack fraction system

+++ server/rooms.py
@@ -306,8 +306,8 @@ def load_dungeon_templates(directory: str = "rooms/dungeon1", type_id: str = "d1
                         "dialog": dlg, "personality": pers,
                     })
                 elif tokens[0] == "monsters" and len(tokens) >= 3:
-                    # Dynamic group: monsters <kind> <base_count>
-                    monster_groups.append({"kind": tokens[1], "count": int(tokens[2])})
+                    # Dynamic group: monsters <kind> <pack_fraction>
+                    monster_groups.append({"kind": tokens[1], "count": float(tokens[2])})
                 elif tokens[0] == "monster" and len(tokens) >= 4:
                     monsters.append({"kind": tokens[1], "x": int(tokens[2]), "y": int(tokens[3])})
